+++ ownlang-core/src/main/java/com/annimon/ownlang/util/input/InputSourceDetector.java
@@ -3,15 +3,19 @@ package com.annimon.ownlang.util.input;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-public record InputSourceDetector() {
+public record InputSourceDetector(String basePath) {
     public static final String RESOURCE_PREFIX = "resource:";
+
+    public InputSourceDetector() {
+        this("");
+    }
 
     public boolean isReadable(String programPath) {
         if (programPath.startsWith(RESOURCE_PREFIX)) {
             String path = programPath.substring(RESOURCE_PREFIX.length());
-            return getClass().getResource(path) != null;
+            return getClass().getResource(basePath + path) != null;
         } else {
-            Path path = Path.of(programPath);
+            Path path = Path.of(basePath, programPath);
             return Files.isReadable(path) && Files.isRegularFile(path);
         }
     }
@@ -19,9 +23,9 @@ public record InputSourceDetector() {
     public InputSource toInputSource(String programPath) {
         if (programPath.startsWith(RESOURCE_PREFIX)) {
             String path = programPath.substring(RESOURCE_PREFIX.length());
-            return new InputSourceResource(path);
+            return new InputSourceResource(basePath + path);
         } else {
-            return new InputSourceFile(programPath);
+            return new InputSourceFile(basePath + programPath);
         }
     }
 }