+++ ownlang-parser/src/main/java/com/annimon/ownlang/parser/ParserMetadata.java
@@ -1,0 +1,6 @@
+package com.annimon.ownlang.parser;
+
+public record ParserMetadata(
+        String basePath
+) {
+}